Backend modularization: Move project and task routes to blueprints
- Moved all project-related endpoints to backend/routes/projects.py as a Flask blueprint (projects_bp)
- Moved all task-related endpoints to backend/routes/tasks_routes.py as a Flask blueprint (tasks_bp)
- Registered new blueprints in app.py and removed legacy inline route definitions
- Refactored app.py to clean up duplicate app/blueprint registration and ensure only one app instance
- Updated utils.py with all task helper/validation functions required by the new blueprints
- All backend tests pass for modularized project and task routes
- Checklist updated to reflect completed and in-progress backend modularization steps
